requirements.txt
fastapi
sqlmodel
uvicorn
pyJWT
bcrypt
passlib
pytest
